Split parts relation routes
