+++ backend/tests/test_audit_log_coverage.py
@@ -344,28 +344,28 @@ def test_each_mutator_writes_audit_row(
         ("parts_core.create_part", _setup_part_create, "part.created", "part", _created_part_id),
         ("parts_core.patch_part", _setup_part_patch, "part.updated", "part", _target_id),
         (
-            "parts_core.add_substitute",
+            "parts_relations.add_substitute",
             _setup_part_substitute_add,
             "part.substitute_added",
             "part_substitute",
             _target_ids,
         ),
         (
-            "parts_core.del_substitute",
+            "parts_relations.del_substitute",
             _setup_part_substitute_delete,
             "part.substitute_removed",
             "part_substitute",
             _target_ids,
         ),
         (
-            "parts_core.add_member",
+            "parts_relations.add_member",
             _setup_part_member_add,
             "part.member_added",
             "part_meta_member",
             _target_ids,
         ),
         (
-            "parts_core.del_member",
+            "parts_relations.del_member",
             _setup_part_member_delete,
             "part.member_removed",
             "part_meta_member",
